New directories added - log and middleware; Middleware logic removed from index.js
